fix: correct import paths after restructuring
This commit corrects the import paths in the application after the project restructuring.
